requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.4.0
python-multipart